requirements.txt
fastmcp>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
markdown>=3.5
beautifulsoup4>=4.12.0
watchdog>=3.0.0
